+++ src/lib.rs
@@ -41,6 +41,7 @@ extern crate tokio_io;
 extern crate tokio_reactor;
 extern crate tokio_signal;
 extern crate tokio_tcp;
+extern crate tokio_threadpool;
 extern crate tokio_timer;
 extern crate trust_dns_resolver;