+++ CHANGES.md
@@ -6,8 +6,11 @@
 
 * Use tokio
 
+* `System` refactored to accommodate tokio runtime.
+
 * `Arbiter::handle()` is not available anymore. 
   Use `Arbiter::spawn()` and `Arbiter::spawn_fn()` instead.
+
 
 
 ## 0.5.7 (2018-05-17)